Add '--no-new-instance' option.
This would allow us to force Geany to not open a new instance even
if no file arguments are passed in the command-line.

Some tweaks were also added so that the current instance window would
activate itself even if no new file arguments are specified whenever
this option is used.