Fix and enhance sorting of editor tabs based on pathnames
Simple sorting by comparing pathname strings is not enough, and produces
inconsistent results.  Files need to be grouped and sorted based on
their parent directories.  This update does that, and also imports
get_doc_folder() to notebook.c, and uses it, so that the results would
be exactly the same as the one in documents list.