Give '.cloneX' filenames to cloned documents
The filename is generated by adding a suffix to it in the form of
'.cloneX' where X is a number ranging from 0 to 99.  X is an empty
string if it is 0.

Any generated filename in which the file it targets already exists, or
any generated filename that is already being used by another opened
document, gets skipped, and the iteration moves to the next number.  If
the iteration reaches 99 and no appropriate filename was still
generated, the cloned document is simply given with no filename or NULL.