Get debug mode from the GEANY_DEBUG environment variable
Debug mode can now also be enabled by setting this variable to a
nonempty string.

Additionally, if it set to a valid debug domain, that domain will be
used as a set value to G_MESSAGES_DEBUG instead of "all".

Currently recognized domains are "error", "warning", "critical",
"message", "info", "debug", and the special domain "all".